requirements.txt
backports.ssl-match-hostname==3.4.0.2
dnspython==1.12.0
future==0.15.0
python-dockercloud==1.0.3
python-etcd==0.4.3
pytz==2013.8
requests==2.1.0
six==1.9.0
urllib3==1.14
websocket-client==0.32.0
wsgiref==0.1.2
